requirements.txt
requests==2.32.3
environs==14.1.1
python-telegram-bot==13.7
